Fixed unit tests and added catch_warnings to all tests.
