+++ tests/models/yolo/unit/test_darknet_network.py
@@ -1,9 +1,10 @@
 import warnings
 
 import pytest
+import torch.nn as nn
 from pytorch_lightning.utilities.warnings import PossibleUserWarning
 
-from pl_bolts.models.detection.yolo.yolo_config import (
+from pl_bolts.models.detection.yolo.darknet_network import (
     _create_convolutional,
     _create_maxpool,
     _create_shortcut,
@@ -40,13 +41,14 @@ def test_create_convolutional(config, catch_warnings):
         assert conv.conv.padding == (pad_size, pad_size)
 
     if config["batch_normalize"]:
-        assert len(conv) == 3
+        assert isinstance(conv.norm, nn.BatchNorm2d)
 
-    if activation != "linear":
-        if activation != "logistic":
-            assert activation == conv[-1].__class__.__name__.lower()[: len(activation)]
-        elif activation == "logistic":
-            assert "sigmoid" == conv[-1].__class__.__name__.lower()
+    if activation == "linear":
+        assert isinstance(conv.act, nn.Identity)
+    elif activation == "logistic":
+        assert isinstance(conv.act, nn.Sigmoid)
+    else:
+        assert conv.act.__class__.__name__.lower().startswith(activation)
 
 
 @pytest.mark.parametrize(
@@ -73,12 +75,18 @@ def test_create_maxpool(config, catch_warnings):
         category=PossibleUserWarning,
     )
 
-    pad_size = (config["size"] - 1) // 2
+    #    print("size", config["size"])
+    #    print("stride", )
+    pad_size, remainder = divmod(max(config["size"], config["stride"]) - config["stride"], 2)
+    #    print("pad_size", pad_size)
     maxpool, _ = _create_maxpool(config, [3])
+    #    print("maxpool.maxpool.padding", maxpool.maxpool.padding)
 
-    assert maxpool.kernel_size == config["size"]
-    assert maxpool.stride == config["stride"]
-    assert maxpool.padding == pad_size
+    assert maxpool.maxpool.kernel_size == config["size"]
+    assert maxpool.maxpool.stride == config["stride"]
+    assert maxpool.maxpool.padding == pad_size
+    if remainder != 0:
+        assert isinstance(maxpool.pad, nn.ZeroPad2d)
 
 
 @pytest.mark.parametrize(
